LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/cors-crossorigin-requests-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: SyntaxError: JSON Parse error: Expected '}'
json-module-crossorigin



PASS Imported JSON module, cross-origin with CORS
PASS Imported JSON module, cross-origin, missing CORS ACAO header
FAIL Imported JSON module with parse error, cross-origin, with CORS assert_equals: Unexpected _log value expected "SyntaxError" but got "0-0"
